Add some Kconfig options for cflash
Make cflash build if SCSI is enabled, and CXL is enabled.

There's probably a chance that we need to change / tune it to be a
built-in if cxl is built-in and module if cxl is a module. For now,
to get this to work, build cxl and cflash as built-ins (make sure
the .config for your build has "=y" for CFLASH and CXL_BASE.