cflash_disk_attach implementation
This drop adds the disk_attach ioctl implementation. This
has serious issues in that the file descriptor returned
cannot me used for mmap etc. as the filp_open on a
character special file is not the equivalent of chrdev_open.

Also moved the ioctl structures to a separate file
(cflash_ioctl.h)

Stripped out much of the legacy initialization element
that were carried over from user space.