Resolve afu_sync (alignment) problem
This drop resolves the issue with afu_sync(). The problem was
that the first command (AFU_INIT_INDEX) was aligned but not
the others, especially AFU_SYNC_INDEX.

Also wire the scsi_dev->hostdata to carry the luninfo structure
and use them in the ioctls. This is a move towards multi-lun
support.