Resolve ordering issue with cxl_attach_fd
This drop gets the first set of read/write from user
application to work. The issue earlier was that
the user context's capabilities were getting reset
by the AFU on the cxl_attach_fd() call.

This drop also rounds up the size so that we
allocate in chunk multiples.

Make the lun ID real, though it looks awkard in LE format.

Removed the NEWCXL ifdef.