Remove the hard-coded WWPN
Use the kernel services provided to retrieve the VPD.
Integrate it with the provisiong parser from user space
to get at the WWPN.
Program the WWPN into the AFU.
So this drop essentially should get the driver functional
on any system.

Also call the pci_enable_device() service since that
is also now functional.