Private buffers per command, check out/in cmds
This drop introduces private buffers per command, so
that multiple operations can be done concurrently from
the SCSI stack. The locking does not need to be
as strict with private buffers. Also introduced
services to check out a command block and release
it after the command has completed.

Moved hexdump() to cflash_util.c