Make the parameters consistent
This drop merges some of the Surelock parameters and gets
it consistent with the other SCSI parameters.

Each port on the CORSA adapter is really a *bus*.
Each bus has only one target (since we are directly attached to the
	Texan)
Each target can have a number of LUNs (TBD).

Make the lunidarray dynamically allocated.