Update physical LUN support
This drop fixes the format 1 RHT entries such that the driver
is in sync with the AFU and populating the entry in a format
that the AFU understands. A dummy format 1 structure is used
to correctly fill out the lower-half of the RHT entry and then
update the entry in a single write operation.

Also included are various fixes/changes:
- transitioned the initialization of a command to occur within get_next_cmd()
- commented out the port mask branch for internal lun as it requires 0x3 (for now)
- made calls to memset/memcpy consisten (don't use &foo[0] for arrays)
- made lfd and work (members of the lun_info struct) into arrays
- temporarily reduced the number of supported LUNs to 8 (with the new arrays came
  a larger memory footprint and caused us to fail _get_free_pages(), Matt will
  revert with a larger change set to rework how the lun data is organized/allocated)

Change-Id: I628927556bd35d39d063bfcc6ff025ae71afd35e