Transition lun_info to a list
This drop completes the first wave of work to reduce the AFU
memory footprint. The lun_info structures are now allocated as
needed and managed in a list.

Also included is a 2-line change to support SG DMA. For now, the
original lines have been left commented out for comparison while
we continue to investigate integration with the Linux SCSI stack.

Change-Id: I5a0daac01d9cffa63e7daaebef3f23d4bfb27f42