Fix up port_sel and lun_id in cflash_send_scsi
This drop also limits the number of buses and
the number of targets per bus.

Allows the kernel/master context to send read/write commands.

Uses the pci_set_dma_mask service now that the interface
is functional.

Get started on passing SCSI, FC and AFU errors back.

This drop is not sufficient to deprecate the REPORT LUNS
(find_lun).