Add read_cap16 to disk_attach
With this drop, we should be fully operational with the scsi stack
scanning for us (qc=1). The legacy scaffolding remains while we finish
working out any last details.

We are unable to determine the disk capactiy when using the scsi
stack to scan. Therefore we will send a read_cap16 to the device
on the very first disk_attach ioctl that comes down for that device.

This drop also changes the default return code for our ioctl handler
to be -EINVAL instead of -EOPNOTSUPP as that is what other handlers
seem to return and the upper stack may not behave the same without it.

Note, ba_terminate() is safe to call unconditionally. We may want to make
ba_init() similar if we're concerned about recursive paths due to the new
scanning method.

Change-Id: I2fcbf77b4390f9efbfbd967dd96137d46b3adb5a