Add cflash_dev_* macros
Similar to the simplification of cflash_* trace macros, this drop adds
wrappers to automatically add the function name and newline character
to entries that use the dev_* trace facilities.

This drop removes the ENTER/LEAVE macros as they were no longer used
or required. Adjustments were also made to the trace strings and comments
in cflash_ba.c to shorten them to 80 character lines.

Change-Id: If5534340d1a7a43237c82d800e727cfc14dcffb9