Increase size of command pool
This drop doubles the size of the command pool to bypass a
corner case that exists due to how the last entry is reserved
for the AFU sync command. Round-robin logic is also introduced
in the command checkout routine to encourage better performance
during concurrent or rapid sequential operations.

Two additional bug fixes in the interrupt handler are included:

 - the buffer must be unmapped prior to calling scsi_done()
 - commands must be checked back in regardless of if they had
   a scsi command 'hidden' within them