Add support for device specific values
This driver could support multiple devices, each with unique
values for certain functions. With this change, a structure
is introduced to hold these values on a per-device basis and
bind them via the PCI device id table. During probe, they can
be associated with the device and values (that are known to be
supported with that device) referenced/used.

For this initial drop, only max_sectors exists at this time.