Save LUN access permissions on DK_CAPI_ATTACH
Users specify their desired read/write access permissions
to be used for all accesses made via the context they're
attaching to during the the DK_CAPI_ATTACH ioctl. This
drop saves the translated permissions (passed in as O_*
flags) within the internal context structure such that
they may be used when programming entries in the RHT.

Included is also a small bug fix in DK_CAPI_CLONE to
allow continued execution even when a load from mbox_r
returns a value of 0.