Remove context dup and mbox_r challenge concepts
The notion of allowing contexts to point to the same RHT
was a stale holdover from GA1 and is no longer being
pursued. An additional holdover that is no longer needed
is the validation dance that mbox_w/mbox_r support to
'serialize' access between threads and the master context.

Both of these concepts have been removed entirely. The
mbox_r is still read immediately before writing to the
ctx_cap register as it is architecturally required to
'unlock' the register.