Bug fix for error path
If a context was closed in cflash_disk_attach (say due to
error returned from cflash_afu_attach), do not expect to
be able to look up a valid context during last release.
This used to cause an oops in misc/cxl, but now that
the release is trapped, it triggers an oops in cflash.

Make sure that context_id is an int(signed), since that is
what is returned from cxl_process_element().