Minor cleanup in prepration for RFC drop
This drop includes:
Elimination of compilation warnings.
Get the PCI init and tear down functions to be more complete
Change the Adapter name so that it is consistently reported
Bump up the revision number/date.
Add MODULE_DEVICE_TABLE entry so that the driver is loaded
automatically when the VPHB detects the AFU.