cflash: Add RHT-LUN mapping table
Add support for a table that maps LUNs to resource handles that were
created for the LUN. This is required for backtracking when freeing
up resources on a non-user-initiated teardown. It also serves to
validate resource handles against a specific LUN, a requirement when
operating in an environment that allows multiple LUNs to be accessed
via the same context.