cflash: Link close handler to detach
Add support for cleaning up a context on a close of the adapter
file descriptor. This includes support for performing a close on
behalf of the user during the final detach for a given context.
Note that temporarily, this is only performed when the newly
introduced module parameter, close, is set to 1. This was done
as there are users who are checking the return code of close on
the adapter file descriptor and might see an EBADF failure due
to the close now being performed on detach.

Also added is a pair of sanity asserts (BUG_ON) for the user
counter of a LUN to ensure they don't go negative. These are
intended to help catch bugs during the development cycle.