cflash: Perform close on clone
The duplicated adapter file descriptor for the source context needs to be
closed on a clone to prevent stale entries from lingering in the fd table
for the process. The user will need to explicitly perform this close for
any contexts not cloned.

Add code to copy over the lun access list from the source to destination
context. Don't enforce the user having previously called reuse on the
destination context. Don't bother with trying to clean up this list on
memory allocation failure.