cflash: Use tgid instead of pid
Linux uses the pid member of the task structure to track the process id
but also to track the lightweight process id when a process is multithreaded.

This causes an issue when trying to bookkeep contexts to pids. Instead, the
tgid field maintains the pid across all threads and only changes when fork
is called and the child process obtains a new pid (tgid).