cflash: Check some rc's on term_mc
Some of Mikey's API's that previously were void now can return an
error. Check for these return codes and assert if they are in error,
indicating a sequencing bug.

Note, leave commented out until we have updated to Mikey's latest API,
as the compile will fail with an error when trying to interpret a void
function return code.