cflash: Add support for mmap error notification
Create an mmap handler that 'catches' mmap faults from the user application
and either calls into the CXL fault handler or 'resolves' the fault using
a dummy page of all foxes. An error-aware user application can then determine
if the underlying infrastructure is in error context by looking for all foxes
on read.

Create a routine to unmap a previously established mapping to support toggling
between the error page and the actual mapping.

Create a secret ioctl to test the toggle until we're properly plugged in to the
error recovery framework.