cflash: Remmove volatiles keyword for all MMIO
MMIO is performed through accessors (ie: readq/writeq) that already guarantee
that operations will not be optimized away by the compiler. Leave the hardware
queue's current pointer defined as volatile for now as it points to a location
that is shared with the AFU and accessed directly (not protected by an
accessor wrapper).