cflash: Lock down host_use fields
Use the slock embedded with each command to serialize updates of the
host_use_b[0] field which is used to track the state of the command.

Also simplify paths where used and combine clearing of this field with
the clearing of the retry field (host_use_b[1]) to occur during command
checkin. Commands should not be referenced after checkin is performed
on them.