cxl: Allow the kernel to trust that an image won't change on PERST.
Provide a kernel API and a sysfs entry which allow a user to specify
that when a card is PERSTed, it's image will stay the same, allowing
it to participate in EEH.

cxl_reset is used to reflash the card. In that case, we cannot safely
assert that the image will not change. Therefore, disallow cxl_reset
if the flag is set.

Signed-off-by: Daniel Axtens <dja@axtens.net>

Conflicts:
	drivers/misc/cxl/api.c
	drivers/misc/cxl/cxl.h
	drivers/misc/cxl/sysfs.c
	include/misc/cxl.h