cflash: Introduce context error table
Use a secondary table to maintain a snapshot of the context table when
in an error recovery state. When an error occurs, use the mark_context_error
routine to snapshot and then clear the context table. Each active context
is then updated to reflect that error recovery is required and it's mapping
removed to notify the user (when they fault again) that the context is in
an error state.