cflash: Move to own TMF spinlock
Borrowing the TMF waitq's spinlock seemd to cause a stall condition when
waiting for the TMF to complete. To rememdy, introduce our own spin lock
to serialize TMF and use the appropriate wait services.

Also add a timeout while waiting for a TMF completion. When a TMF times
out, report back a failure such that a bigger hammer reset can occur.

Fix the timeout for AFU commands to be represented in jiffies to run
and not a future jiffies clock value (when we moved over from a timer
this was not changed and causes the timeouts to be insanely large).

Trace the AFU command address as returned on a checkout for easier
tracking in the code.