cflash: Add NULL fops for error handling
When calls to CXL services fail in disk_attach(), the error handling
code cleans up work that was already performed. Part of this work
includes 'putting back' the file structure that was allocated. This is
performed via the fput() service, which in turn calls the release fop
associated with the file. Doing so triggers a call into the CXL services
when they are not expecting one, leading to a crash. By overriding the
fops right before calling fput(), the release fop is NULL'ed out, solving
the problem.