cxl: allocate and release the SPA with the AFU
Previously the SPA was allocated and freed upon entering and leaving
AFU-directed mode. This causes some issues for error recovery - contexts
hold a pointer inside the SPA, and they may persist after the AFU has
been detached.

As such, allocate the SPA when the AFU is allocated, and don't release
it until the AFU is released.

Signed-off-by: Daniel Axtens <dja@axtens.net>