cxl: rework adapter link checks.
We need to guard every time cxl_whatever_(read|write) is called. This
is because a call to those functions will dereference an offset into an
mmio register, and the mmio mappings get invalidated in the EEH
teardown.

Previously, this was done with multiple little checks.
Move the check into the read/write functions in the header.
This gives them the same semantics as usual PCI operations:
 - a write to a channel that is down is ignored.
 - a read from a channel that is down returns all fs.

This allows us to refine a number of the checks in native.c.
 - Be more quiet if there's an error in a teardown path
 - Throw a big WARN if someone tries to start a CXL operation
   while the card is down.

Signed-off-by: Daniel Axtens <dja@axtens.net>