cflash: Write LUN table dynamically
All previous versions wrote the LUN table as it was
being built (either during scan/config time or during
the manage_lun ioctl). Starting with this version the
LUN table will only be written on the first open in virtual
mode for a LUN. Which also means that the LUN table is not
used for physical LUNs. In addition the LUN table will
only grow to as many virtual LUNs that are being used,
not as many that are available.