cflash: Wait for EEH recovery in context_recovery
If/when an AFU_RECOVER ioctl slips in just before an EEH notification,
the MMIO error detection logic is executed and detects the failure. In
this scenario, the application expects that the context is recovered
upon return. Thus, when this occurs, wait on the eeh recovery and then
attempt to recover.

Note that the time before the eeh recovery occurs and this ioctl being
invoked may take some time (up to a second or just over). To avoid
looping around and around during this window, sleep for 1 second
between the time the MMIO failure is detected and the time the wait
is entered.