cflash: Notify users via error page during remove
When an adapter is to be removed, running users need to be notified and
given a chance to shut down their contexts. Move stop_term_user_contexts
to be called from cxlflash_remove() and update the routine to use the
error page notification mechanism to notify users and wait until they
have quiesced.

Related bug fixes:
- remove a context from the error list when it is removed
- while waiting for a context to be freed during the error marking,
use the correct running threads count and sleep in 1 second intervals
instead of busy-waiting
- once marking contexts in error is complete, schedule other work on
the current CPU to give threads a chance to respond before proceeding
further