cflash: Conditional error list removal
On context tear down, the context may or may not be in the error list,
and the list deletion service does not protect against a node not being
in the list (crash occurs). Add a check prior to calling list_del() to
ensure that the context was linked in to avoid a crash.