cflash: Bypass AFU sync when in failed EEH state
AFU sync operations should be gated during EEH recovery. When a recovery
fails and an adapter is to be removed, sync requests can occur as part of
cleaning up resources associated with an adapter prior to its removal. In
this scenario, these requests are identified here and simply ignored (safe
due to the AFU going away).