cflash: Use list_del_init() when removing context
When taking a context out of the error list and placing it
back in the table upon error recovery, use the list_del_init()
service to unlink the context from the error recovery list. This
will reinitialize the context list node entry such that it can
be safely used with list_empty() on a detach operation.