cflash: Clear context control mapping on error
When transitioning a context to the error state, clear (NULL) it's
control mapping as it will no longer be valid with the AFU reset.

In destroy_context(), evaluate the existence of the context's control
map instead of the AFU's mapping before proceeding with the MMIO writes
to clear the RHT registers and capabilities.