cflash: Lock down contexts
Introduce per-context mutexes that serialize an entire chain of
execution for a context. All resources within the context are
protected via the per-context mutex. The mutex is taken in get_context()
and held for the useful lifetime of the context reference. In areas
where contexts lookup were previously performed multiple times during
the useful lifetime of the context reference, functions have been changed
to support passing a context reference to avoid having to drop the mutex
or attempt a nested mutex acquisition (a no-no).

The context table spinlock has been changed to a mutex as it is not used
in an interrupt environment. This mutex protects the per-adapter context
table and error recovery list.

By changing to a mutex for the context table/list, the per-context nref is
no longer needed and had been removed. A retry dance located in the
get_context() routine serializes between threads wishing to acquire a
context and an already in-flight context removal thread.