cflash: Don't AFU sync contexts on err recov list
When contexts are on the error recovery list, they were moved there
as part of an AFU teardown operation (either a reset or a removal
cleanup). When in this state, they are unknown to the AFU and
therefore an AFU sync should not be performed. Trying to sync might
cause an inconsistent AFU state and if performed during a reset,
could crash due to the AFU's MMIO space being reset.

This commit avoids performing an AFU sync on teardown paths for
contexts on the error list (err_recovery_active is TRUE). Other
paths don't need to check as they are properly fenced by the
get_context() management bits (can't find contexts on the error
recovery list).