cflash: Check rc of scsi_normalize_sense()
The scsi_normalize_sense() routine indicates if the sense data was found
and valid via a boolean return code. Use this to evaluate if the code
should continue to process the provided sense data.

Fix a multiplication bug where 3 32-bit values are multiplied and
overflow into a 64-bit quantity. The Linux compiler masks the
upper bits, losing the overflow which breaks subsequent math.

Add better tracing in the verify ioctl path.