cxl: Don't cxl_reset when perst_reloads_same_image
cxl_reset is designed to be used to reflash the card.

perst_reloads_same_image asserts that a card's personality does not
change on PERST. If we are reflashing the card, we can't confidently
assert this. Therefore, only allow cxl_reset to be used if the
perst_reloads_same_image flag is not set.