cflash: Avoid context deadlock
The order in which a context is obtained from the table/list must
always be maintained when taking the protecting mutexes. The table/list
mutex must be obtained first followed by the individual context's mutex.
There should almost never be a case where a thread that holds a particular
context's mutex tries to acquire the table/list mutex without first
dropping the context mutex. The one exception is when the context cannot
be found by any other threads. These cases should be well documented.