cflash: Fix remove ordering
The scsi_host_put() routine will cause the host and it's associated memory
to be freed. Therefore, as the inverse function of scsi_host_alloc(), it
needs to be called last in a termination scenario.

Also, use the cancel_workq_sync() instead of flush_workq(), and call earlier
in the teardown sequence to flush out the workq.