cflash: Free block allocator on lun detach
The block allocator is blindly allocated any time a LUN transitions
from MODE_NONE to MODE_VIRTUAL. Therefore, if the block allocator is
not freed when transitioning from MODE_VIRTUAL back to MODE_NONE {when
the last user detaches from the LUN] then there is a leak of memory.

Free the block allocator inside of detach_lun() when moving back to
MODE_NONE, regardless of the previous mode to avoid unnecessary logic
(the block allocator free routine is safe to call this way).