cflash: Move gli->slock to mutex
Use a mutex to protect the global lun info structure instead
of a spin lock. This allows for allocations with simpler code
structure due to not being bound by the constraints of a spin
lock. Also fix a potential memory leak when failing after a
successful init_vlun() but before a successful attach.