cflash: Close remaining hole in ioctl drain
The ioctl drain mechanism relies upon a read semaphore being held across
the execution of the ioctl handler. When taken after the adapter state
is evaluated, a hole exists in a reset scenario that is attempting to
drain ioctls: ioctls that have already passed the state check but not
yet acquired the semaphore are 'overlooked' during the drain operation.
This can result in ioctls racing with a reset/recovery thread that
depends on ioctls being quiesced (recall that new ioctl theads are
blocked [reset_waitq] during a reset operation).

To remedy, widen the scope of read semaphore that covers the ioctl
execution path. Take the semaphore first thing when the main ioctl
handler is invoked and don't release it until execution has completed,
just before returning to the calling stack. Likewise, only drop the
semaphore just before entering a wait condition (reset_waitq) and
immediately reacquire it when exiting the wait condition. Because the
semaphore will be obtained before [and held during] the adapter state
check, the hole no longer exists. The potential for 'overlooked' ioctls
is a thing of the past.