cflash: Fix logic bug and release context
The logic in get_context() that matches a CXL context with the stored
off one in the cxlflash context needs to be updated to handle the case
when falling back to !ctx and !file.

The new event_pending and deliver_event routines needs to release the
context via put_context() after successfully obtaining a good context.