cflash: Always attempt LUN table initialization
On a virtual LUN open, there should always be an attempt made to
update the LUN table for the adapter, regardless of the the global
LUN's state prior to the open. This is necessary in a multi-adapter
environment so that access to a LUN [that has already been transitioned
to virtual mode] through a second (or greater) adapter is configured
correctly by programming the adapter's LUN table.