cxlflash: Obtain additional sdev reference
When a LUN is removed, the sdev that is associted with the LUN remains
intact until its reference count drops to 0. In order to prevent an sdev
from being removed while a context is still associated with it, obtain an
additional reference per-context for each LUN attached to the context.

This resovles a potential Oops in the release handler when a dealing with
a LUN that has already been removed.