cxlflash: Properly obtain AFU version and check it
The AFU version is stored as a non-terminated string of bytes within
a 64-bit little-endian register. Use the correct MMIO accessor to read
this value into a null-terminated buffer and add a check to prevent a
GA1 AFU from being used with this driver as it doesn't support the GA1
interface.