cxlflash: Correct usage of scsi_host_put()
Currently, scsi_host_put() is being called prematurely in the
remove path and is missing entirely in an error cleanup path.
Move the usage on remove to be the last cleanup action taken
and introduce a call to scsi_host_put() in the one initialization
error path that does not use remove to cleanup.